moved test to datasettimeutil file
