+++ frontend/src/data/providers/CawpProvider.ts
@@ -16,6 +16,7 @@ import {
   UNKNOWN_W,
   HISP_W,
 } from '../utils/Constants'
+import { getMostRecentYearAsString } from '../utils/DatasetTimeUtils'
 
 export const CAWP_CONGRESS_COUNTS: MetricId[] = [
   'women_this_race_us_congress_count',
@@ -94,7 +95,7 @@ class CawpProvider extends VariableProvider {
 
     df = this.filterByGeo(df, breakdowns)
 
-    const mostRecentYear = this.getMostRecentYear(df, metricQuery.metricIds, metricQuery.dataTypeId)
+    const mostRecentYear = getMostRecentYearAsString(df, metricQuery.metricIds)
    
     df = this.filterByTimeView(df, timeView, mostRecentYear)
     df = this.renameGeoColumns(df, breakdowns)
