+++ frontend/src/data/utils/DatasetTimeUtils.ts
@@ -1,3 +1,4 @@
+import { IDataFrame } from 'data-forge'
 import {
   type TimeSeries,
   type TrendsData,
@@ -277,4 +278,24 @@ export function getMinMaxGroups(data: TrendsData): DemographicGroup[] {
   ]
 
   return lowestAndHighestGroups
+}
+
+export function getMostRecentYearAsString(df: IDataFrame, metricIds: MetricId[]): string {
+  const metricId = metricIds[0]
+
+  const filteredRows = df
+  .where(row => row?.[metricId] !== undefined)
+  .select(row => ({
+    time_period: row.time_period,
+    metricId: row?.[metricId],
+  }))
+
+  const distinctYearsWithData = filteredRows
+    .select(row => row.time_period)
+    .distinct()
+    .toArray()
+
+  const mostRecentYear = Math.max(...distinctYearsWithData)
+
+  return mostRecentYear.toString()
 }
